+++ java/lib/DumbNumbSetTest.java
@@ -7,13 +7,13 @@ class DumbNumbSetTest {
   public static void main(String[] args) throws Exception {
     int size = 1000000;
     Random r = new Random(6283);
-    HashMap<Long, Boolean> h = new HashMap<Long, Boolean>();
+    HashMap<Integer, Boolean> h = new HashMap<Integer, Boolean>();
     DumbNumbSet ns = new DumbNumbSet();
 
     if(args.length == 1 && args[0].equals("hash")) {
 
       for(int i = 0; i < size; i++) {
-        long n = (long) r.nextInt(size * 10);
+        int n = r.nextInt(size * 10);
         h.put(n, true);
         //h.put((long)i, true);
         if(h.get(n) == null) {
@@ -23,7 +23,7 @@ class DumbNumbSetTest {
     }
     else {
       for(int i = 0; i < size; i++) {
-        long n = (long) r.nextInt(size * 10);
+        int n = r.nextInt(size * 10);
         ns.add(n);
         //ns.add(i);
         if(!ns.hasValue(n)) {
